Fixed router references, added webpack dev config.
