+++ src/components/routes/Blog.vue
@@ -1,0 +1,9 @@
+<template>
+  <h1>Blog</h1>
+</template>
+
+<script>
+  const Blog = Vue.component('Blog', {
+    template: '#blog-template'
+  });
+</script>
