+++ src/components/routes/Home.vue
@@ -1,4 +1,4 @@
-<script type="text/x-template" id="home-template">
+<template>
   <div class="container">
     <h1 v-localize="'header.nav.home'"></h1>
     <button @click="$setLang('en')">English</button>
@@ -7,12 +7,9 @@
     <button @click="store.commit('status')" class="button button-primary">Change loading status</button>
     <author fname="John Nolette" image="app/assets/images/author.jpg" github="https://github.com/neetVeritas" twitter="https://twitter.com/neet_jn"></author>
   </div>
-</script>
+</template>
 
-<style>
-</style>
-
-<script type="text/javascript">
+<script>
   const Home = Vue.component('Home', {
     template: '#home-template'
   });
