--- src/components/routes/blog.vue
@@ -1,8 +1,0 @@
-<script type="text/x-template" id="blog-template">
-  <h1>Blog</h1>
-</script>
-<script type="text/javascript">
-  const Blog = Vue.component('Blog', {
-    template: '#blog-template'
-  });
-</script>
